requirements.txt
faiss-cpu>=1.10.0
langchain>=0.3.17
langchain-community>=0.3.16
langchain-ollama>=0.2.3
litellm>=1.60.2
openai>=1.61.0
openpyxl>=3.1.5
pandas>=2.2.3
pypdf2>=3.0.1
python-docx>=1.1.2
python-dotenv>=1.0.1
python-pptx>=1.0.2
streamlit==1.41.1








